Fix some compiler warnings

+++ libraries/Utils/Timer.cpp
@@ -32,7 +32,7 @@ void Timer::stop() {
 void Timer::drawGuiWindow(GLFWwindow* window) {
     ImGui::SetNextWindowSize(ImVec2(300, 100), ImGuiSetCond_FirstUseEver);
     ImGui::Begin("Performance");
-    ImGui::PlotLines("Frametime", m_ftimes.data(), m_ftimes.size(), 0, nullptr, 0.0f, std::numeric_limits<float>::max());
+    ImGui::PlotLines("Frametime", m_ftimes.data(), static_cast<int>(m_ftimes.size()), 0, nullptr, 0.0f, std::numeric_limits<float>::max());
     float flaccTime = 0.0f;
     if (m_ftimes.size() > 21) {
         for (size_t i = m_ftimes.size() - 21; i < m_ftimes.size(); ++i) {
